refactor: 각 스크린에 맞는 함수 파일 재배치
