+++ app/src/main/java/com/sopt/now/compose/Sources.kt
@@ -20,53 +20,6 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 
-@Composable
-fun SignInTextField(
-    value: String,
-    onValueChange: (String) -> Unit,
-    label: String,
-    leadingIcon: ImageVector,
-    modifier: Modifier = Modifier,
-    isPassword: Boolean = false,
-) {
-    TextField(
-        value = value,
-        onValueChange = onValueChange,
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(10.dp),
-        label = { Text(label) },
-        leadingIcon = { Icon(leadingIcon, contentDescription = null) },
-        singleLine = true,
-        visualTransformation = if (isPassword) PasswordVisualTransformation() else VisualTransformation.None
-    )
-}
-
-@Composable
-fun SignUpTextField(
-    value: String,
-    onValueChange: (String) -> Unit,
-    label: String,
-    leadingIcon: ImageVector?,
-    modifier: Modifier = Modifier,
-    isPassword: Boolean = false
-) {
-    TextField(
-        value = value,
-        onValueChange = onValueChange,
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(10.dp),
-        label = { Text(label) },
-        leadingIcon = if (leadingIcon != null) {
-            { Icon(leadingIcon, contentDescription = null) }
-        } else {
-            null
-        },
-        singleLine = true,
-        visualTransformation = if (isPassword) PasswordVisualTransformation() else VisualTransformation.None
-    )
-}
 
 @Composable
 fun SOPTOutlinedButton(text: Int, onClick: () -> Unit, enabled: Boolean) {
@@ -75,77 +28,6 @@ fun SOPTOutlinedButton(text: Int, onClick: () -> Unit, enabled: Boolean) {
     ) {
         Text(text = stringResource(id = text))
     }
-}
-
-fun isSignInValid(
-    navController: NavController,
-    context: Context,
-    inputID: String,
-    inputPW: String,
-    ID: String,
-    PW: String,
-    Name: String,
-    Place: String
-) {
-    when {
-        (inputID.isEmpty()) -> {
-            showToast(context, R.string.toast_SignIn_InvalidSignIn_IDBlank)
-        }
-
-        (inputPW.isEmpty()) -> {
-            showToast(context, R.string.toast_SignIn_InvalidSignIn_PWBlank)
-        }
-
-        (inputID == ID && inputPW == PW) -> {
-            showToast(context, R.string.toast_SignIn_ValidSignIn)
-            navController.navigate("MyPage?ID=$inputID&PW=$inputPW&Name=$Name&Place=$Place")
-        }
-
-        (inputID != ID) -> {
-            showToast(context, R.string.toast_SignIn_InvalidID)
-        }
-
-        (inputPW != PW) -> {
-            showToast(context, R.string.toast_SignIn_InvalidPW)
-        }
-    }
-}
-
-fun isSignUpValid(
-    navController: NavController,
-    context: Context,
-    ID: String,
-    IDValid: Boolean,
-    PW: String,
-    PWValid: Boolean,
-    Name: String,
-    NameValid: Boolean,
-    Place: String
-) {
-    if (ID.isEmpty()||PW.isEmpty()||
-        Name.isEmpty()||Place.isEmpty())
-        showToast(context, R.string.toast_SignUp_InvalidSignUp_Blank)
-
-    else if (IDValid && PWValid && NameValid) {
-        showToast(context, R.string.toast_SignUp_ValidSignUp)
-        navController.navigate("SignIn?ID=$ID&PW=$PW&Name=$Name&Place=$Place")
-    }
-
-    else
-        showToast(context,R.string.toast_SignUp_InvalidSignUp)
-}
-
-@Composable
-fun MyPageText(
-    text: String
-) {
-    Text(
-        text = text,
-        modifier = Modifier.padding(10.dp),
-        fontSize = 20.sp,
-        fontWeight = FontWeight.Bold,
-        textAlign = TextAlign.Center
-    )
 }
 
 fun showToast(context: Context, message: Int) {
