+++ app/src/main/java/com/sopt/now/compose/screen/MyPageScreen.kt
@@ -6,13 +6,16 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.sopt.now.compose.MyPageText
+import androidx.compose.ui.unit.sp
 import com.sopt.now.compose.R
 
 @Composable
@@ -41,4 +44,17 @@ fun MyPageScreen(ID: String, PW: String, Name: String, Place: String) {
         MyPageText(text = stringResource(R.string.txt_MyPage_Place))
         MyPageText(text = Place)
     }
+}
+
+@Composable
+fun MyPageText(
+    text: String
+) {
+    Text(
+        text = text,
+        modifier = Modifier.padding(10.dp),
+        fontSize = 20.sp,
+        fontWeight = FontWeight.Bold,
+        textAlign = TextAlign.Center
+    )
 }
